% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/atminv.R
\docType{package}
\name{atminv}
\alias{atminv}
\alias{atminv-package}
\title{atminv}
\description{
Reproducible package for Section 4.1 of "Spatio-temporal bivariate statistical models for atmospheric trace-gas inversion"
}

