% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{EM}
\alias{EM}
\title{Laplace functions for the ATCP problem}
\usage{
EM(s_obs, C_m, Qobs, B, S_f_log, mu_f_log, Yf_thresh, Y_init, theta_init,
  ind = 1:ncol(B), s_mol = NULL, t_mol = NULL, n_EM = 100,
  model = "full")
}
\description{
EM method needs input
s_obs -- observation df
C_m -- C matrix relating Z to MF
Qobs -- observation precision
B -- atmospheric model matrix
corr_t -- temporal correlation function with one free parameter
corr_s -- spatial correlation function with one free parameter
S_f_log -- covariance function of log(Yf)
mu_f_log -- exepectation of log(Yf)
Yf_thresh -- threshold at which to start dropping Yf nodes
Y_init -- initial Y values
theta_init -- initial theta values
n_EM -- number of EM iterations
}

