% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{ggplot_dual_axis}
\alias{ggplot_dual_axis}
\title{ggplot-dual-axis}
\usage{
ggplot_dual_axis(lhs, rhs, axis.title.y.rhs = "rotate")
}
\description{
See http://stackoverflow.com/questions/18989001/how-can-i-put-a-transformed-scale-on-the-right-side-of-a-ggplot2
}

