% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoxCox.R
\name{log_f_theta_f}
\alias{log_f_theta_f}
\title{Conditional for flux field parameters}
\usage{
log_f_theta_f(theta, Z, D, flux_cov_fn, X, scale_ul = 2, nu_ul = 2,
  lambda_ll = -3, lambda_ul = 3, lambda_fix = NA, uncorrelated = FALSE,
  Pericchi = TRUE)
}
\arguments{
\item{theta}{flux field parameters -- see details}

\item{Z}{matrix of flux field data (of size m x n_obs, where n_obs is generally 2; an inventory and the sample derived from the mole fraction observations)}

\item{D}{distance matrix of flux field spatial locations}

\item{flux_cov_fn}{covariance function used to generate the transformed correlation matrix. This should be of the form flux_cov_fn(D,scale,nu), where scale and nu are scale and smoothness parameters.}

\item{X}{covariates matrix (of size n_f x p)}

\item{scale_ul}{upper limit of uniform distribution over scale parameter. Default of 2}

\item{nu_ul}{upper limit of uniform distribution over smoothness parameter. Default of 2}

\item{lambda_fix}{use a fixed value for lambda (do not sample from it). If this is set then \code{theta} should only contain two numbers}

\item{lambda_l}{absolute limit of uniform distribution over Box-Cox parameter. Default of 3}
}
\description{
Returns the conditional distribution of the flux field parameters given the flux field and, possibly, other flux maps. Uniform priors are assumed over the scale and smoothness, the bounds of which can be specified.
}
\details{
To be completed
}
\examples{
n_f <- 20
s <- 1:n_f
cov_fn <- function(D,scale,nu) {
     exp(-scale*(D^nu))    
}
print(log_f_theta_f(theta = c(1,1,0),
              Z = matrix(rpois(n_f,100)),
              D = as.matrix(dist(s)),
              flux_cov_fn = cov_fn,
              X = matrix(rnorm(n_f),n_f,1)))
}

