% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BoxCox.R
\name{log_f_theta_m}
\alias{log_f_theta_m}
\title{Conditional for mole-fraction field discrepancy parameters}
\usage{
log_f_theta_m(theta, Yf, s, t, C_m, Qobs, Z, B, Q_zeta_fn)
}
\arguments{
\item{theta}{mole-fraction field parameters -- see details}

\item{s}{observation spatial locations}

\item{t}{observation temporal indices}

\item{C_m}{observation incidence matrix (of size m x n_m)}

\item{Qobs}{observation precision matrix (of size m x m)}

\item{Z}{matrix of mole fraction observations (of size m x 1)}

\item{B}{source-receptor relationship matrix (of size n_m x n_f)}

\item{Q_zeta_fn}{function which returns the inverse of the mole-fraction discrepancy correlation matrix. This function should be of the form \code{ Q_zeta_fn(s,t,d_t,d_s)} where \code{s,t} are as above, \code{d_t} is the temporal length scale parameter and \code{d_s} is the spatial length scale parameter}
}
\description{
Returns the conditional distribution of the mole-fraction field discrepancy parameters given the mole-fraction observations and the flux field
}
\details{
To be completed
}
\examples{
n_m <- m_obs <- 100
n_f <- 20
Yf <- matrix(rpois(n_f,20),n_f,1)
log_f_theta_m( theta = c(4,0.8,1),
               Yf=Yf,
               s = 1:10,
               t = 1:10,
               C_m = .symDiagonal(n_m),
               Qobs = Diagonal(n_m),
               Z = 200 + matrix(rnorm(n_m,sd=100),n_m,1),
               B = matrix(rnorm(n_m*n_f),n_m,n_f),
               Q_zeta_fn = function(s,t,d_t,d_s) Diagonal(length(t) * length(s)))
}

