% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GMRF_classes.R
\docType{class}
\name{Obs-class}
\alias{Obs-class}
\title{observation block}
\description{
An observation block inherits from class \code{block}, however can only be used in certain ways in a graph. For example it can be linked to an observation
process block but not to another observation block. To initialise use \code{initObs}.
}
\details{

}

