% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoxCox.R
\name{Yf_BC_conditionals}
\alias{Yf_BC_conditionals}
\title{Conditional for Yf}
\usage{
Yf_BC_conditionals(Z, Zinv, C_m, Qobs, B, X, Q_zeta, S_f_trans, lambda = 0,
  lambda_fix = NA, Pericchi = TRUE, ind = 1:nrow(S_f_trans))
}
\arguments{
\item{Z}{matrix of mole fraction observations (of size m x 1)}

\item{C_m}{observation incidence matrix (of size m x n_m)}

\item{Qobs}{observation precision matrix (of size m x m)}

\item{B}{source-receptor relationship matrix (of size n_m x n_f)}

\item{X}{covariates matrix (of size n_f x p)}

\item{Q_zeta}{precision matrix of discrepancy field (of size n_m x n_m). Should be sparse}

\item{S_f_trans}{covariance matrix of flux field in transformed (Gaussian) space (of size n_f x n_f)}

\item{lambda}{Box-Cox transformation parameter}

\item{ind}{vector of indices on which to consider the flux (some unidentifiable grid cells may be omitted)}
}
\description{
Returns the conditional distribution and its derivative of the flux field given the flux field parameters and the mole-fraction observations
}
\examples{
n_f <- 20
n_m <- m_obs <- 100
cond_fns <- Yf_BC_conditionals(Z = 200 + matrix(rnorm(n_m,sd=100),n_m,1),
                               C_m = .symDiagonal(n_m),
                               Qobs = Diagonal(n_m),
                               B = matrix(rnorm(n_m*n_f),n_m,n_f),
                               X = matrix(rep(1,n_f)),
                               Q_zeta = n_m*Diagonal(n_m),
                               S_f_trans = Diagonal(n_f),
                               lambda = 0.1
                             )
Yf <- matrix(rpois(n_f,20),n_f,1)                              
cond_fns$logf(Yf)
cond_fns$gr_logf(Yf)
}

