% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMRF_classes.R
\docType{class}
\name{GMRF_AR-class}
\alias{GMRF_AR-class}
\title{@docType class
 @title GMRF_AR}
\description{
An autoregressive proces which inherits from class \code{GMRF}. The primary difference is that this class is constructed using
a first-order auto-regressive structure. Unlike the random walk, this is not an intrinsic GMRF.
}

